requirements.txt
anyio==3.7.1
arabic-reshaper==3.0.0
asgiref==3.7.2
attrs==25.4.0
bcrypt==4.1.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2025.10.5
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
deep-translator==1.11.4
distlib==0.3.7
Django==5.0
dnspython==2.6.1
email-validator==2.1.0.post1
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.46.0
graphviz==0.20.1
greenlet==3.0.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
node==1.2.1
npm==0.1.1
numpy==1.26.4
odict==1.9.0
openpyxl==3.1.2
openwa==1.3.16
optional-django==0.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
platformdirs==3.11.0
plumber==1.7
protobuf==4.21.12
psycopg2-binary==2.9.9
pycparser==2.23
PyMySQL==1.1.1
pyodbc==5.1.0
pyparsing==3.1.1
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.8.2
python-pptx==0.6.23
pytz==2023.3.post1
reportlab==4.4.3
requests==2.32.3
rfc3986==1.5.0
scipy==1.11.4
selenium==4.36.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.4.4
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.5.0
virtualenv==20.24.6
websocket-client==1.9.0
Werkzeug==3.1.1
wsproto==1.2.0
WTForms==3.2.1
xlrd==2.0.1
XlsxWriter==3.1.9
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.1
zope.lifecycleevent==5.0
zope.proxy==5.1
Flask
gunicorn
Werkzeug
deep-translator==1.11.4
openai>=1.0.0
APScheduler==3.10.4

